refactor: 점수 계산 위치 이동 (JanggiGame)
